refactor: Explode the add dataset dialog into smaller files
