+++ app/configurator/components/add-dataset-dialog/add-dataset-dialog.tsx
@@ -6,7 +6,6 @@ import {
   Checkbox,
   CircularProgress,
   Collapse,
-  DialogActions,
   DialogContent,
   DialogProps,
   DialogTitle,
@@ -15,33 +14,18 @@ import {
   IconButton,
   IconButtonProps,
   InputAdornment,
-  Link,
   ListItemText,
   MenuItem,
   OutlinedInput,
-  paperClasses,
   Select,
   SelectChangeEvent,
-  Stack,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableRow,
   TextField,
-  Tooltip,
   Typography,
-  TypographyProps,
   useEventCallback,
 } from "@mui/material";
-import { Theme } from "@mui/material/styles";
-import { makeStyles } from "@mui/styles";
-import clsx from "clsx";
-import groupBy from "lodash/groupBy";
 import keyBy from "lodash/keyBy";
-import maxBy from "lodash/maxBy";
 import uniq from "lodash/uniq";
-import { FormEvent, useCallback, useEffect, useMemo, useState } from "react";
+import { FormEvent, useEffect, useMemo, useState } from "react";
 import { useClient } from "urql";
 
 import {
@@ -51,14 +35,22 @@ import {
   SearchDatasetResultsCount,
   SearchDatasetSortControl,
 } from "@/browser/dataset-browse";
-import { FirstTenRowsCaption } from "@/browser/dataset-preview";
 import { getEnabledChartTypes } from "@/charts";
 import Flex from "@/components/flex";
-import { Error as ErrorHint, HintOrange, Loading } from "@/components/hint";
 import Tag from "@/components/tag";
-import { ConfiguratorStateConfiguringChart, DataSource } from "@/config-types";
+import { ConfiguratorStateConfiguringChart } from "@/config-types";
 import { getChartConfig } from "@/config-utils";
 import { RightDrawer } from "@/configurator/components/drawers";
+import {
+  CautionAlert,
+  useCautionAlert,
+} from "@/configurator/components/add-dataset-dialog/caution-alert";
+import {
+  inferJoinBy,
+  JoinBy,
+} from "@/configurator/components/add-dataset-dialog/infer-join-by";
+import PreviewDataTable from "@/configurator/components/add-dataset-dialog/preview-table";
+import { SearchOptions } from "@/configurator/components/add-dataset-dialog/types";
 import {
   addDatasetInConfig,
   isConfiguring,
@@ -67,26 +59,17 @@ import {
 import {
   ComponentTermsets,
   Dimension,
-  isJoinByComponent,
-  isStandardErrorDimension,
   isTemporalDimensionWithTimeUnit,
-  Measure,
-  Termset,
 } from "@/domain/data";
 import { truthy } from "@/domain/types";
 import {
   executeDataCubesComponentsQuery,
   useDataCubesComponentsQuery,
-  useDataCubesMetadataQuery,
-  useDataCubesObservationsQuery,
 } from "@/graphql/hooks";
-import { isJoinById, joinDimensions } from "@/graphql/join";
 import { ComponentId } from "@/graphql/make-component-id";
 import {
-  DataCubeComponentsQuery,
   SearchCubeFilterType,
   SearchCubeResultOrder,
-  useDataCubeComponentsQuery,
   useDataCubeComponentTermsetsQuery,
   useSearchCubesQuery,
 } from "@/graphql/query-hooks";
@@ -98,7 +81,8 @@ import { Locale } from "@/locales/locales";
 import { useLocale } from "@/locales/use-locale";
 import { useEventEmitter } from "@/utils/eventEmitter";
 import useEvent from "@/utils/use-event";
-import useLocalState from "@/utils/use-local-state";
+
+import useStyles from "./use-styles";
 
 const DialogCloseButton = (props: IconButtonProps) => {
   return (
@@ -115,503 +99,6 @@ const DialogCloseButton = (props: IconButtonProps) => {
     </IconButton>
   );
 };
-
-const useStyles = makeStyles((theme: Theme) => ({
-  addButton: {
-    transition: "opacity 0.25s ease",
-  },
-  dialogCloseArea: {
-    position: "absolute",
-    top: "1rem",
-    right: "1rem",
-    display: "flex",
-    gap: "0.5rem",
-    alignItems: "center",
-  },
-  newAnnotation: {
-    color: theme.palette.success.main,
-  },
-  listTag: {
-    whiteSpace: "break-spaces",
-  },
-  listItemSecondary: {
-    display: "flex",
-    flexWrap: "wrap",
-    alignItems: "center",
-    gap: 1,
-    paddingTop: theme.spacing(1),
-  },
-  previewSelect: {
-    width: 302,
-  },
-  previewSelectPaper: {
-    [`& .${paperClasses.root}`]: {
-      marginTop: "0.25rem",
-    },
-  },
-}));
-
-const NewAnnotation = (props: TypographyProps) => {
-  const classes = useStyles();
-  return (
-    <Typography
-      className={clsx(classes.newAnnotation, props.className)}
-      lineHeight={1}
-      variant="caption"
-      {...props}
-    >
-      <Trans id="dataset.search.preview.new-dimension">New</Trans>
-    </Typography>
-  );
-};
-
-const useCautionAlert = () => {
-  const [isOpen, setIsOpen] = useLocalState("add-dataset-caution-alert", true);
-
-  return {
-    isOpen,
-    open: useCallback(() => setIsOpen(true), [setIsOpen]),
-    close: useCallback(() => setIsOpen(false), [setIsOpen]),
-  };
-};
-
-const CautionAlert = ({ onConfirm }: { onConfirm: () => void }) => {
-  return (
-    <HintOrange>
-      <Trans id="dataset.search.caution.body">
-        The linking of different datasets carries risks such as data
-        inconsistencies, scalability issues, and unexpected correlations. Be
-        sure to use to merge datasets only when you are confident that data
-        should be merged together.
-      </Trans>
-      <Box sx={{ mt: 1 }}>
-        <Link
-          onClick={(ev) => {
-            ev.preventDefault();
-            onConfirm();
-          }}
-          href="#"
-        >
-          <Trans id="dataset.search.caution.acknowledge">
-            Understood, I&apos;ll proceed cautiously.
-          </Trans>
-        </Link>
-      </Box>
-    </HintOrange>
-  );
-};
-
-const inferJoinBy = (
-  leftOptions: SearchOptions[],
-  rightCube: PartialSearchCube
-) => {
-  const possibleJoinBys = leftOptions
-    .map((leftOption) => {
-      const type = leftOption.type;
-      // For every selected dimension, we need to find the corresponding dimension on the other cube
-      switch (type) {
-        case "temporal":
-          return {
-            // TODO Do not only use the first originalIds
-            left:
-              isJoinById(leftOption.id) && leftOption.originalIds
-                ? leftOption.originalIds?.[0].dimensionId
-                : leftOption.id,
-            right: rightCube?.dimensions?.find(
-              (d) =>
-                // TODO Find out why this is necessary
-                d.timeUnit ===
-                `http://www.w3.org/2006/time#unit${leftOption.timeUnit}`
-            )?.id,
-          };
-        case "shared":
-          return {
-            left: [leftOption.id],
-            right: rightCube?.dimensions?.find((d) =>
-              d.termsets.some((t) =>
-                leftOption.termsets.map((t) => t.iri).includes(t.iri)
-              )
-            )?.id,
-          };
-        default:
-          const exhaustiveCheck: never = type;
-          return exhaustiveCheck;
-      }
-    })
-    .filter(
-      (x): x is { left: ComponentId; right: ComponentId } =>
-        !!(x.left && x.right)
-    );
-
-  return possibleJoinBys.reduce<{
-    left: string[];
-    right: string[];
-  }>(
-    (acc, item) => {
-      acc.left.push(item.left);
-      acc.right.push(item.right);
-      return acc;
-    },
-    { left: [], right: [] }
-  );
-};
-
-type JoinBy = ReturnType<typeof inferJoinBy>;
-
-const PreviewDataTable = ({
-  dataSource,
-  existingCubes,
-  currentComponents,
-  searchDimensionsSelected,
-  otherCube,
-  onClickBack,
-  onConfirm,
-  addingDataset,
-}: {
-  dataSource: DataSource;
-  existingCubes: { iri: string }[];
-  currentComponents:
-    | DataCubeComponentsQuery["dataCubeComponents"]
-    | undefined
-    | null;
-  searchDimensionsSelected: SearchOptions[];
-  otherCube: PartialSearchCube;
-  onClickBack: () => void;
-  onConfirm: () => void;
-  addingDataset: boolean;
-}) => {
-  const locale = useLocale();
-  const commonQueryVariables = {
-    sourceType: dataSource.type,
-    sourceUrl: dataSource.url,
-    locale,
-  };
-
-  const inferredJoinBy = useMemo(
-    () => inferJoinBy(searchDimensionsSelected, otherCube),
-    [searchDimensionsSelected, otherCube]
-  );
-
-  const [otherCubeComponentsQuery] = useDataCubeComponentsQuery({
-    variables: {
-      ...commonQueryVariables,
-      cubeFilter: { iri: otherCube.iri },
-    },
-    pause: !otherCube,
-  });
-
-  const otherCubeComponents = otherCubeComponentsQuery.data?.dataCubeComponents;
-
-  const isFetchingComponents = otherCubeComponentsQuery.fetching;
-
-  const isQueryPaused = !otherCubeComponents || !currentComponents;
-
-  const cubeFilters = [
-    {
-      iri: currentComponents!.dimensions?.[0].cubeIri,
-      joinBy: inferredJoinBy.left,
-    },
-    {
-      iri: otherCube.iri,
-      joinBy: inferredJoinBy.right,
-    },
-  ];
-
-  console.log(cubeFilters);
-  const [observations] = useDataCubesObservationsQuery({
-    pause: isQueryPaused,
-    variables: {
-      locale,
-      sourceType: dataSource.type,
-      sourceUrl: dataSource.url,
-      cubeFilters,
-    },
-  });
-  const [currentCubes] = useDataCubesMetadataQuery({
-    variables: {
-      sourceType: dataSource.type,
-      sourceUrl: dataSource.url,
-      locale: locale,
-      cubeFilters: existingCubes.map((cube) => ({ iri: cube.iri })),
-    },
-  });
-
-  const isFetching =
-    isFetchingComponents || observations.fetching || currentCubes.fetching;
-
-  const allColumns = useMemo(() => {
-    const shouldIncludeColumnInPreview = (d: Dimension | Measure) =>
-      !isStandardErrorDimension(d);
-    const currentDimensions = (currentComponents?.dimensions ?? []).filter(
-      (x) => shouldIncludeColumnInPreview(x)
-    );
-    const currentMeasures = (currentComponents?.measures ?? []).filter((x) =>
-      shouldIncludeColumnInPreview(x)
-    );
-    const otherDimensions = (otherCubeComponents?.dimensions ?? []).filter(
-      (x) => shouldIncludeColumnInPreview(x)
-    );
-    const otherMeasures = (otherCubeComponents?.measures ?? []).filter((x) =>
-      shouldIncludeColumnInPreview(x)
-    );
-
-    const joinedColumns = joinDimensions([
-      {
-        dataCubeComponents: {
-          dimensions: currentDimensions,
-          measures: [],
-        },
-        joinBy: inferredJoinBy.left,
-      },
-      {
-        dataCubeComponents: {
-          dimensions: otherDimensions,
-          measures: [],
-        },
-        joinBy: inferredJoinBy.right,
-      },
-    ]);
-
-    const {
-      join: joinedJoinDimensions = [],
-      other: joinedOtherDimensions = [],
-      current: joinedCurrentDimensions = [],
-    } = groupBy(joinedColumns, (d) => {
-      const isJoinBy = isJoinByComponent(d);
-      return isJoinBy
-        ? "join"
-        : d.cubeIri === otherCube.iri
-          ? "other"
-          : "current";
-    });
-
-    return [
-      ...joinedJoinDimensions,
-      ...joinedOtherDimensions,
-      ...otherMeasures,
-      ...joinedCurrentDimensions,
-      ...currentMeasures,
-    ];
-  }, [
-    currentComponents?.dimensions,
-    currentComponents?.measures,
-    inferredJoinBy.left,
-    inferredJoinBy.right,
-    otherCube.iri,
-    otherCubeComponents?.dimensions,
-    otherCubeComponents?.measures,
-  ]);
-
-  const [selectedColumnsRaw, setSelectedColumns] = useState<
-    undefined | string[]
-  >(undefined);
-  useEffect(() => {
-    if (
-      otherCubeComponents?.dimensions &&
-      otherCubeComponents.measures &&
-      selectedColumnsRaw === undefined
-    ) {
-      setSelectedColumns(allColumns.map((x) => x.id));
-    }
-  }, [
-    allColumns,
-    otherCubeComponents?.dimensions,
-    otherCubeComponents?.measures,
-    selectedColumnsRaw,
-  ]);
-  const selectedColumns = useMemo(
-    () => selectedColumnsRaw ?? [],
-    [selectedColumnsRaw]
-  );
-
-  const selectedColumnsById = useMemo(
-    () => Object.fromEntries(selectedColumns.map((x) => [x, true])),
-    [selectedColumns]
-  );
-
-  const previewObservations = useMemo(() => {
-    const data = observations.data?.dataCubesObservations.data ?? [];
-    const bestObservation = maxBy(data, (obs) => {
-      return allColumns.reduce((acc, dim) => acc + (obs[dim.id] ? 1 : 0), 0);
-    });
-
-    const amount = 8;
-    const index = bestObservation ? data.indexOf(bestObservation) : 0;
-    const clampedIndex = Math.max(0, Math.min(index, data.length - amount));
-    return data.slice(clampedIndex, clampedIndex + amount);
-  }, [allColumns, observations.data?.dataCubesObservations.data]);
-
-  return (
-    <>
-      <DialogContent sx={{ overflowX: "hidden" }}>
-        <Box sx={{ display: "flex", flexDirection: "column", gap: "1rem" }}>
-          <Typography variant="h2">
-            <Trans id="dataset.search.preview.title">
-              Review available dimensions
-            </Trans>
-          </Typography>
-          <Typography variant="body1">
-            <Trans id="dataset.search.preview.description">
-              Review all available dimensions before continuing to edit your
-              visualization.
-            </Trans>
-          </Typography>
-          <div>
-            <Typography variant="h6" mb={1}>
-              <Trans id="dataset.search.preview.datasets">Datasets</Trans>
-            </Typography>
-            <Stack direction="column" spacing={1}>
-              <Typography variant="caption">
-                {currentCubes.data?.dataCubesMetadata
-                  .map((metadata) => metadata.title)
-                  .join(", ")}
-              </Typography>
-              <div>
-                <NewAnnotation />
-                <br />
-                <Typography variant="caption" mt={-1} component="div">
-                  {otherCube.title}
-                </Typography>
-              </div>
-            </Stack>
-          </div>
-          {isFetching ? <Loading delayMs={0} /> : null}
-          {observations.error ? (
-            <ErrorHint>
-              <Box
-                component="details"
-                sx={{ width: "100%", cursor: "pointer" }}
-              >
-                <summary>{observations.error.name}</summary>
-                {observations.error.message}
-              </Box>
-            </ErrorHint>
-          ) : null}
-        </Box>
-        {otherCubeComponents ? (
-          <>
-            {observations.data?.dataCubesObservations ? (
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  alignItems: "flex-end",
-                  gap: 2,
-                }}
-              >
-                <Box
-                  sx={{
-                    overflowX: "scroll",
-                    width: "100%",
-                    mt: 6,
-                    mb: 4,
-                    boxShadow: 4,
-                  }}
-                >
-                  <Table>
-                    <TableHead>
-                      <TableRow>
-                        {allColumns.map((column) =>
-                          !!selectedColumnsById[column.id] ? (
-                            <TableCell
-                              key={column.id}
-                              sx={{ minWidth: 200, maxWidth: 300 }}
-                            >
-                              {column.cubeIri === otherCube.iri && (
-                                <NewAnnotation />
-                              )}
-                              {isJoinByComponent(column) ? (
-                                <>
-                                  <Tooltip
-                                    arrow
-                                    title={
-                                      <>
-                                        {column.originalIds
-                                          .map((o) => o.description)
-                                          .join(", ")}
-                                      </>
-                                    }
-                                    placement="right"
-                                  >
-                                    <Tag type="dimension">Joined</Tag>
-                                  </Tooltip>
-                                </>
-                              ) : null}
-                              <br />
-                              <Typography variant="h5">
-                                {column.label}
-                              </Typography>
-                            </TableCell>
-                          ) : null
-                        )}
-                      </TableRow>
-                    </TableHead>
-                    <TableBody>
-                      {previewObservations.map((observation, index) => (
-                        <TableRow key={index}>
-                          {allColumns.map((column) =>
-                            !!selectedColumnsById[column.id] ? (
-                              <TableCell key={column.id}>
-                                {observation[column.id]}
-                              </TableCell>
-                            ) : null
-                          )}
-                        </TableRow>
-                      ))}
-                    </TableBody>
-                  </Table>
-                </Box>
-                <FirstTenRowsCaption />
-              </Box>
-            ) : null}
-          </>
-        ) : null}
-      </DialogContent>
-      <DialogActions
-        sx={{
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "flex-end",
-          gap: 1,
-          px: 4,
-          pt: "0 !important",
-          pb: "2rem !important",
-        }}
-      >
-        <Box sx={{ display: "flex", gap: 2 }}>
-          <Button variant="outlined" onClick={onClickBack}>
-            <Trans id="button.back">Back</Trans>
-          </Button>
-          <LoadingButton
-            loading={addingDataset}
-            variant="contained"
-            onClick={onConfirm}
-          >
-            <Trans id="button.confirm">Confirm</Trans>
-          </LoadingButton>
-        </Box>
-      </DialogActions>
-    </>
-  );
-};
-
-type SearchOptions =
-  | {
-      type: "temporal";
-
-      /** Contains for now the join by id */
-      id: ComponentId;
-      label: string;
-      timeUnit: string;
-      originalIds: Dimension["originalIds"];
-    }
-  | {
-      type: "shared";
-      /** Technically it's an iri, but we keep the id name for the sake of type consistency. */
-      id: string;
-      label: string;
-      termsets: Termset[];
-    };
 
 const extractSearchDimensions = (
   dimensions: Dimension[],
@@ -733,12 +220,6 @@ export const DatasetDialog = ({
     SearchCubeResultOrder.Score
   );
   const [includeDrafts, setIncludeDrafts] = useState(false);
-
-  const commonQueryVariables = {
-    sourceType: state.dataSource.type,
-    sourceUrl: state.dataSource.url,
-    locale,
-  };
 
   const activeChartConfig = state.chartConfigs.find(
     (chartConfig) => chartConfig.key === state.activeChartKey
@@ -871,13 +352,22 @@ export const DatasetDialog = ({
   const { isOpen, open, close } = useCautionAlert();
 
   const ee = useEventEmitter();
+
+  const inferredJoinBy = useMemo(
+    () =>
+      otherCube
+        ? inferJoinBy(selectedSearchDimensions ?? [], otherCube)
+        : undefined,
+    [selectedSearchDimensions, otherCube]
+  );
+
   const handleConfirm = useEventCallback(async () => {
-    if (!components || !otherCube) {
+    if (!components || !otherCube || !inferredJoinBy) {
       return null;
     }
 
     await addDataset({
-      joinBy: inferJoinBy(selectedSearchDimensions ?? [], otherCube),
+      joinBy: inferredJoinBy,
       otherCube: otherCube,
     });
     handleClose({}, "escapeKeyDown");
@@ -1075,13 +565,14 @@ export const DatasetDialog = ({
           </DialogContent>
         </>
       )}
-      {otherCube ? (
+      {otherCube && inferredJoinBy ? (
         <PreviewDataTable
           key={otherCube.iri}
           dataSource={state.dataSource}
           currentComponents={components}
           existingCubes={currentCubes}
           otherCube={otherCube}
+          inferredJoinBy={inferredJoinBy}
           addingDataset={addingDataset}
           onConfirm={handleConfirm}
           onClickBack={() => setOtherCube(undefined)}
